+++ triplea/service/repository/export/engine/__init__.py
@@ -4,17 +4,27 @@ import click
 from triplea.schemas.article import Article
 from triplea.utils.general import print_error
 from triplea.service.click_logger import logger
+from triplea.config.settings import SETTINGS
 
-
-def export_engine(fx_filter, fx_transform, fx_output, limit_sample=0, proccess_bar=True):
+def export_engine(fx_filter,
+                  fx_transform,
+                  fx_output,
+                  limit_sample=0,
+                  proccess_bar=True):
     l_id = PERSIST.get_article_id_list_by_cstate(1, "FlagShortReviewByLLM") # TODO Change it
+    SETTINGS.AAA_CLI_ALERT_POINT
     n = 0
     filter_number = 0
     doc_number = len(l_id)
     if doc_number == 0:
         return
     if proccess_bar:
-        bar = click.progressbar(length=doc_number, show_pos=True, show_percent=True)
+        bar = click.progressbar(length=doc_number,
+                                show_pos=True,
+                                show_percent=True)
+    else:
+        logger.INFO("Start export...")
+
 
     output_list = []
     for id in l_id:
@@ -39,8 +49,12 @@ def export_engine(fx_filter, fx_transform, fx_output, limit_sample=0, proccess_b
 
             # For View Proccess
             if proccess_bar:
-                bar.label = f"""{filter_number} Article(s) Updated."""
+                bar.label = f"""{filter_number} Article(s) exported."""
                 bar.update(1)
+            else:
+                if n % SETTINGS.AAA_CLI_ALERT_POINT == 0:
+                    logger.INFO(f"{n} Article(s) exported.")
+
 
             if limit_sample != 0:  # Unlimited
                 if n > limit_sample:
@@ -49,5 +63,7 @@ def export_engine(fx_filter, fx_transform, fx_output, limit_sample=0, proccess_b
             print()
             print(logger.ERROR(f"article. ID = {id}"))
             print_error()
-
+            
+    if proccess_bar is False:
+        logger.INFO("End export.")
     return output_list